[sourcegen] Reshuffle content to reduce imports
