+++ interfaces/sourcegen/sourcegen/_helpers.py
@@ -5,7 +5,6 @@ from dataclasses import fields
 import inspect
 from pathlib import Path
 import textwrap
-import re
 try:
     from ruamel import yaml
 except ImportError:
@@ -13,6 +12,7 @@ except ImportError:
 
 
 def read_config(config_file: Path) -> dict:
+    """Read YAML configuration file."""
     if config_file.is_file():
         with config_file.open("r", encoding="utf-8") as config_file:
             reader = yaml.YAML(typ="safe")
@@ -47,22 +47,3 @@ def normalize_indent(code: str) -> str:
 
 def get_preamble() -> str:
     return Path(__file__).parent.joinpath("preamble.txt").read_text("utf-8")
-
-
-def xml_tag(tag: str, text: str, suffix: str="", index=0) -> str:
-    """Extract content enclosed between XML tags, optionally skipping matches."""
-    if suffix:
-        suffix = f" {suffix.strip()}"
-    regex = re.compile(rf'(?<=<{tag}{suffix}>)(.*?)(?=</{tag}>)')
-    match = re.findall(regex, text)
-    if index >= len(match):
-        return ""  # not enough matches found
-    return match[index]
-
-
-def split_arglist(arglist: str) -> tuple:
-    """Split C++ argument list into text within parentheses and suffix."""
-    arglist = arglist.strip()
-    suffix = arglist[arglist.rfind(")") + 1:]
-    arglist = re.findall(re.compile(r'(?<=\().*(?=\))'), arglist)[0]
-    return arglist, suffix.strip()
